+++ src/main/java/db/TurnDao.java
@@ -1,4 +1,4 @@
-package domain.dao;
+package db;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
